소개팅 상대 조건 설정 기능 추가

+++ src/main/java/com/TingTing/dto/ConditionResponseDto.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class ConditionResponseDto {
-    private int conditionId;
+    private Long conditionId;
 }
