+++ src/main/java/com/TingTing/entity/Conditions.java
@@ -20,7 +20,7 @@ public class Conditions {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "conditions_id", nullable = false)
-    private int conditionsId;
+    private Long conditionsId;
 
     @Column(name = "ai_name", nullable = false, length = 100)
     private String aiName;
